# staff/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: Hello World App
host: 'localhost:3000'
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: X-API-Key
  JWT:
    description: 'Authorization: Bearer {token}'
    type: apiKey
    name: Authorization
    in: header
security:
  - APIKeyHeader: []
    JWT: []
paths:
...
  /orgs:
    post:
      summary: 创建单个组织机构
      description: 创建单个组织机构...
      parameters:
        - in: body
          name: body
          description: 模型参数
          required: true
          schema:
            $ref: '#/definitions/Org'
      responses:
        '200':
          description: 创建成功
    get:
      summary: 查询多个组织机构
      description: 查询多个组织机构...
      parameters:
        - in: query
          name: pageNum
          type: integer
          description: 页码
          default: 1
        - in: query
          name: pageSize
          type: integer
          description: 分页大小
          default: 10
        - in: query
          name: sorter
          type: string
          description: 排序
      responses:
        '200':
          description: 查询成功
          schema:
            type: object
            properties:
              count:
                type: integer
              rows:
                type: array
                items:
                  $ref: '#/definitions/Org'
    patch:
      summary: 更新多个组织机构
      description: 更新多个组织机构...
      parameters:
        - in: body
          name: body
          description: body参数
          schema:
            type: array
            items:
              $ref: '#/definitions/Org'
      responses:
        '200':
          description: 更新成功
    delete:
      summary: 删除多个组织机构
      description: 删除多个组织机构...
      parameters:
        - in: body
          name: body
          description: body参数
          schema:
            type: object
            example:
              id:
                - '1'
                - '2'
      responses:
        '200':
          description: 更新成功
  '/orgs/{id}':
    get:
      summary: 查询单个组织机构
      description: 查询单个组织机构..
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: ID
      responses:
        '200':
          description: 查询成功
          schema:
            $ref: '#/definitions/Org'
    patch:
      summary: 修改单个组织机构
      description: 修改单个组织机构..
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: ID
        - in: body
          name: body
          description: 模型参数
          required: true
          schema:
            $ref: '#/definitions/Org'
      responses:
        '200':
          description: 修改成功
    delete:
      summary: 删除单个组织机构
      description: 删除单个组织机构..
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: ID
      responses:
        '200':
          description: 删除成功
  /orgs/multiple:
    post:
      summary: 创建多个组织机构
      description: 创建多个组织机构...
      parameters:
        - in: body
          name: body
          description: body参数
          schema:
            type: array
            items:
              $ref: '#/definitions/Org'
      responses:
        '200':
          description: 创建成功
  /org:
    get:
      summary: 查询单个组织机构
      description: 查询单个组织机构...
      parameters:
        - in: query
          name: id
          type: string
          description: 不限于id，可以是任意参数
          required: true
      responses:
        '200':
          description: 查询成功
          schema:
            $ref: '#/definitions/Org'
    post:
      summary: 查找或创建单个组织机构
      description: 查找或创建单个组织机构...
      parameters:
        - in: body
          name: body
          description: 模型参数
          required: true
          schema:
            $ref: '#/definitions/Org'
      responses:
        '200':
          description: 创建成功
          schema:
            $ref: '#/definitions/Org'
...
  /resources:
    post:
      summary: 创建单个资源
      description: 创建单个资源...
      parameters:
        - in: body
          name: body
          description: 模型参数
          required: true
          schema:
            $ref: '#/definitions/Resource'
      responses:
        '200':
          description: 创建成功
    get:
      summary: 查询多个资源
      description: 查询多个资源...
      parameters:
        - in: query
          name: pageNum
          type: integer
          description: 页码
          default: 1
        - in: query
          name: pageSize
          type: integer
          description: 分页大小
          default: 10
        - in: query
          name: sorter
          type: string
          description: 排序
      responses:
        '200':
          description: 查询成功
          schema:
            type: object
            properties:
              count:
                type: integer
              rows:
                type: array
                items:
                  $ref: '#/definitions/Resource'
    patch:
      summary: 更新多个资源
      description: 更新多个资源...
      parameters:
        - in: body
          name: body
          description: body参数
          schema:
            type: array
            items:
              $ref: '#/definitions/Resource'
      responses:
        '200':
          description: 更新成功
    delete:
      summary: 删除多个资源
      description: 删除多个资源...
      parameters:
        - in: body
          name: body
          description: body参数
          schema:
            type: object
            example:
              id:
                - '1'
                - '2'
      responses:
        '200':
          description: 更新成功
  '/resources/{id}':
    get:
      summary: 查询单个资源
      description: 查询单个资源..
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: ID
      responses:
        '200':
          description: 查询成功
          schema:
            $ref: '#/definitions/Resource'
    patch:
      summary: 修改单个资源
      description: 修改单个资源..
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: ID
        - in: body
          name: body
          description: 模型参数
          required: true
          schema:
            $ref: '#/definitions/Resource'
      responses:
        '200':
          description: 修改成功
    delete:
      summary: 删除单个资源
      description: 删除单个资源..
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: ID
      responses:
        '200':
          description: 删除成功
  /resources/multiple:
    post:
      summary: 创建多个资源
      description: 创建多个资源...
      parameters:
        - in: body
          name: body
          description: body参数
          schema:
            type: array
            items:
              $ref: '#/definitions/Resource'
      responses:
        '200':
          description: 创建成功
  /resource:
    get:
      summary: 查询单个资源
      description: 查询单个资源...
      parameters:
        - in: query
          name: id
          type: string
          description: 不限于id，可以是任意参数
          required: true
      responses:
        '200':
          description: 查询成功
          schema:
            $ref: '#/definitions/Resource'
    post:
      summary: 查找或创建单个资源
      description: 查找或创建单个资源...
      parameters:
        - in: body
          name: body
          description: 模型参数
          required: true
          schema:
            $ref: '#/definitions/Resource'
      responses:
        '200':
          description: 创建成功
          schema:
            $ref: '#/definitions/Resource'
  /roles:
    post:
      summary: 创建单个资源
      description: 创建单个资源...
      parameters:
        - in: body
          name: body
          description: 模型参数
          required: true
          schema:
            $ref: '#/definitions/Role'
      responses:
        '200':
          description: 创建成功
    get:
      summary: 查询多个资源
      description: 查询多个资源...
      parameters:
        - in: query
          name: pageNum
          type: integer
          description: 页码
          default: 1
        - in: query
          name: pageSize
          type: integer
          description: 分页大小
          default: 10
        - in: query
          name: sorter
          type: string
          description: 排序
      responses:
        '200':
          description: 查询成功
          schema:
            type: object
            properties:
              count:
                type: integer
              rows:
                type: array
                items:
                  $ref: '#/definitions/Role'
    patch:
      summary: 更新多个资源
      description: 更新多个资源...
      parameters:
        - in: body
          name: body
          description: body参数
          schema:
            type: array
            items:
              $ref: '#/definitions/Role'
      responses:
        '200':
          description: 更新成功
    delete:
      summary: 删除多个资源
      description: 删除多个资源...
      parameters:
        - in: body
          name: body
          description: body参数
          schema:
            type: object
            example:
              id:
                - '1'
                - '2'
      responses:
        '200':
          description: 更新成功
  '/roles/{id}':
    get:
      summary: 查询单个资源
      description: 查询单个资源..
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: ID
      responses:
        '200':
          description: 查询成功
          schema:
            $ref: '#/definitions/Role'
    patch:
      summary: 修改单个资源
      description: 修改单个资源..
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: ID
        - in: body
          name: body
          description: 模型参数
          required: true
          schema:
            $ref: '#/definitions/Role'
      responses:
        '200':
          description: 修改成功
    delete:
      summary: 删除单个资源
      description: 删除单个资源..
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: ID
      responses:
        '200':
          description: 删除成功
  /roles/multiple:
    post:
      summary: 创建多个资源
      description: 创建多个资源...
      parameters:
        - in: body
          name: body
          description: body参数
          schema:
            type: array
            items:
              $ref: '#/definitions/Role'
      responses:
        '200':
          description: 创建成功
  /role:
    get:
      summary: 查询单个资源
      description: 查询单个资源...
      parameters:
        - in: query
          name: id
          type: string
          description: 不限于id，可以是任意参数
          required: true
      responses:
        '200':
          description: 查询成功
          schema:
            $ref: '#/definitions/Role'
    post:
      summary: 查找或创建单个资源
      description: 查找或创建单个资源...
      parameters:
        - in: body
          name: body
          description: 模型参数
          required: true
          schema:
            $ref: '#/definitions/Role'
      responses:
        '200':
          description: 创建成功
          schema:
            $ref: '#/definitions/Role'
  /roleresources:
    post:
      summary: 创建单个角色-资源
      description: 创建单个角色-资源...
      parameters:
        - in: body
          name: body
          description: 模型参数
          required: true
          schema:
            $ref: '#/definitions/RoleResource'
      responses:
        '200':
          description: 创建成功
    get:
      summary: 查询多个角色-资源
      description: 查询多个角色-资源...
      parameters:
        - in: query
          name: pageNum
          type: integer
          description: 页码
          default: 1
        - in: query
          name: pageSize
          type: integer
          description: 分页大小
          default: 10
        - in: query
          name: sorter
          type: string
          description: 排序
      responses:
        '200':
          description: 查询成功
          schema:
            type: object
            properties:
              count:
                type: integer
              rows:
                type: array
                items:
                  $ref: '#/definitions/RoleResource'
    patch:
      summary: 更新多个角色-资源
      description: 更新多个角色-资源...
      parameters:
        - in: body
          name: body
          description: body参数
          schema:
            type: array
            items:
              $ref: '#/definitions/RoleResource'
      responses:
        '200':
          description: 更新成功
    delete:
      summary: 删除多个角色-资源
      description: 删除多个角色-资源...
      parameters:
        - in: body
          name: body
          description: body参数
          schema:
            type: object
            example:
              id:
                - '1'
                - '2'
      responses:
        '200':
          description: 更新成功
  '/roleresources/{id}':
    get:
      summary: 查询单个角色-资源
      description: 查询单个角色-资源..
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: ID
      responses:
        '200':
          description: 查询成功
          schema:
            $ref: '#/definitions/RoleResource'
    patch:
      summary: 修改单个角色-资源
      description: 修改单个角色-资源..
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: ID
        - in: body
          name: body
          description: 模型参数
          required: true
          schema:
            $ref: '#/definitions/RoleResource'
      responses:
        '200':
          description: 修改成功
    delete:
      summary: 删除单个角色-资源
      description: 删除单个角色-资源..
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: ID
      responses:
        '200':
          description: 删除成功
  /roleresources/multiple:
    post:
      summary: 创建多个角色-资源
      description: 创建多个角色-资源...
      parameters:
        - in: body
          name: body
          description: body参数
          schema:
            type: array
            items:
              $ref: '#/definitions/RoleResource'
      responses:
        '200':
          description: 创建成功
  /roleresource:
    get:
      summary: 查询单个角色-资源
      description: 查询单个角色-资源...
      parameters:
        - in: query
          name: id
          type: string
          description: 不限于id，可以是任意参数
          required: true
      responses:
        '200':
          description: 查询成功
          schema:
            $ref: '#/definitions/RoleResource'
    post:
      summary: 查找或创建单个角色-资源
      description: 查找或创建单个角色-资源...
      parameters:
        - in: body
          name: body
          description: 模型参数
          required: true
          schema:
            $ref: '#/definitions/RoleResource'
      responses:
        '200':
          description: 创建成功
          schema:
            $ref: '#/definitions/RoleResource'
...
definitions:
...
  Org:
    type: object
    x-isModel: true
    x-plural: Orgs
    x-tableName: org
    description: 组织机构
    required:
      - id
      - name
    properties:
      id:
        type: string
        format: uuid
        description: ID
        uniqueItems: true
      pid:
        type: string
        format: uuid
        description: 父ID
        x-message: 请输入
      corp_code:
        type: string
        format: string(64)
        description: 企业代码
        x-message: 请输入
        x-showTable: true
      name:
        type: string
        format: string(64)
        description: 姓名
        x-message: 请输入
        x-showTable: true
    example:
      id: '123'
      name: testname
...
  Resource:
    type: object
    x-isModel: true
    x-plural: Resources
    x-tableName: resource
    description: 资源
    required:
      - id
      - resource_code
      - name
    properties:
      id:
        type: string
        format: uuid
        description: ID
        uniqueItems: true
      pid:
        type: string
        format: uuid
        description: 父ID
        x-message: 请输入
      resource_code:
        type: string
        format: string(64)
        description: 资源代码
        x-message: 请输入
        x-showTable: true
        uniqueItems: true
      name:
        type: string
        format: string(64)
        description: 资源名称
        x-message: 请输入
        x-showTable: true
    example:
      id: '123'
      name: testname
  Role:
    type: object
    x-isModel: true
    x-plural: Roles
    x-tableName: role
    description: 资源
    required:
      - id
      - name
    properties:
      id:
        type: string
        format: uuid
        description: ID
        uniqueItems: true
      name:
        type: string
        format: string(64)
        description: 角色名称
        x-message: 请输入
        x-showTable: true
    example:
      id: '123'
      name: testname
  RoleResource:
    type: object
    x-isModel: true
    x-plural: RoleResources
    x-tableName: roleresource
    description: 角色-资源
    required:
      - id
      - role_id
      - resource_id
    properties:
      id:
        type: string
        format: uuid
        description: ID
        uniqueItems: true
      role_id:
        type: string
        format: uuid
        description: 角色ID
        x-message: 请输入
        x-showTable: true
      resource_id:
        type: string
        format: uuid
        description: 资源ID
        x-message: 请输入
        x-showTable: true
      action:
        type: integer
        format: int4
        enum:
          - 0
          - 1
          - 2
          - 3
          - 4
        default: 0
        x-message: 请选择
        x-showTable: true
        x-showFilter: true
        x-enumMap:
          '0': ALL
          '1': 增
          '2': 删
          '3': 改
          '4': 查
        x-description: 操作
        description: |
          "操作"
            * 0 - ALL
            * 1 - 增
            * 2 - 删
            * 3 - 改
            * 4 - 查
    example:
      id: '123'
      name: testname
...